+++ src/main/java/gifterz/textme/domain/oauth/config/HttpInterfaceConfig.java
@@ -1,6 +1,7 @@
-package gifterz.textme.domain.oauth.kakao.config;
+package gifterz.textme.domain.oauth.config;
 
-import gifterz.textme.domain.oauth.kakao.controller.KakaoApi;
+import gifterz.textme.domain.oauth.infra.dku.controller.DkuApi;
+import gifterz.textme.domain.oauth.infra.kakao.controller.KakaoApi;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.reactive.function.client.WebClient;
@@ -18,5 +19,15 @@ public class HttpInterfaceConfig {
                 .build();
 
         return httpServiceProxyFactory.createClient(KakaoApi.class);
+    }
+
+    @Bean
+    DkuApi dkuApi() {
+        WebClient client = WebClient.create();
+        HttpServiceProxyFactory httpServiceProxyFactory = HttpServiceProxyFactory
+                .builder(WebClientAdapter.forClient(client))
+                .build();
+
+        return httpServiceProxyFactory.createClient(DkuApi.class);
     }
 }
