+++ src/main/java/gifterz/textme/domain/oauth/entity/AuthType.java
@@ -6,6 +6,7 @@ public enum AuthType {
 
     PASSWORD,
     KAKAO,
+    DKU,
     ;
 
     public static AuthType fromName(String name) {
